Implemented functionality for user contact mapping

+++ controllers/contactController.js
@@ -2,15 +2,15 @@ const asyncHandler = require("express-async-handler"); //USED TO CAPTURE ERROR U
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) =>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json({ contacts });
 });
 
 //@desc create new contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContacts = asyncHandler(async (req, res) =>{
     console.log("The request body is : ", req.body);
     const { name, email, phone } = req.body;
@@ -21,14 +21,15 @@ const createContacts = asyncHandler(async (req, res) =>{
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id,
     });
     res.status(201).json(contact);
 });
 
 //@desc get contact by id
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) =>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -40,13 +41,19 @@ const getContact = asyncHandler(async (req, res) =>{
 
 //@desc update contact by id
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) =>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user's contact");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -58,13 +65,19 @@ const updateContact = asyncHandler(async (req, res) =>{
 
 //@desc Delete contact by id
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) =>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
     }
+    
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user's contact");
+    }
+
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact);
 });
